Split config data model and config loader
