--- src/docfx/config/ConfigConstants.cs
@@ -1,14 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace Microsoft.Docs.Build
-{
-    internal class ConfigConstants
-    {
-        internal const string Files = "files";
-        internal const string Exclude = "exclude";
-        internal const string Extend = "extend";
-        internal const string DefaultLocale = "defaultLocale";
-        internal const string Localization = "localization";
-    }
-}
